Remove another loop (easier with the merge)
